+++ src/gui/component/mod.rs
@@ -1,0 +1,5 @@
+pub mod settings;
+
+// Default values used on multiple elements.
+pub const DEFAULT_FONT_SIZE: u16 = 14;
+pub const DEFAULT_PADDING: u16 = 10;
